+++ platform/node/test/expression.test.js
@@ -1,4 +1,4 @@
-const {run} = require('../../../mapbox-gl-js/test/integration/lib/expression');
+const {run} = require('../../../maplibre-gl-js/test/integration/lib/expression');
 const mbgl = require('../index');
 const ignores = require('./ignores.json');